+++ src/shared/models/cache-anime.go
@@ -1,0 +1,8 @@
+package models
+
+// AnimeCache type used to store cached anime titles
+type AnimeCache struct {
+	ID   string
+	Name string
+	Year uint16
+}
