+++ src/shared/models/cache.go
@@ -1,0 +1,6 @@
+package models
+
+// Cache type used to store cached anime titles
+type Cache struct {
+	anime []AnimeCache
+}
