+++ src/shared/scraper/anime.go
@@ -1,0 +1,5 @@
+package scraper
+
+func List() {
+
+}
