+++ src/v1/models/anime.go
@@ -1,15 +1,5 @@
 package models
 
-// Anime type
-type Anime struct {
-	ID       uint16
-	Name     string
-	AltNames []string
-	Year     uint16
-	Themes   []Theme
-	LinkInfo string
-}
-
 // AnimeEx export type
 type AnimeEx struct {
 	ID       uint16   `json:"id"`
