--- src/v1/utils/cache.go
@@ -1,36 +1,0 @@
-package utils
-
-import (
-	"encoding/json"
-	"io/ioutil"
-	"log"
-	"os"
-	"path/filepath"
-
-	"github.com/eoussama/anusic-api/src/v1/models"
-)
-
-// CachedAnimeList stores the cached anime list
-var CachedAnimeList []models.Anime = []models.Anime{}
-
-// LoadCache loads the cache files
-func LoadCache() bool {
-	log.Println("Loading cache...")
-
-	// Constructing the cache file
-	absPath, _ := filepath.Abs(".")
-	path := filepath.Join(absPath, "data", "cache.json")
-	jsonFile, err := os.Open(path)
-
-	// No cache found
-	if err != nil {
-		return false
-	}
-
-	byteValue, _ := ioutil.ReadAll(jsonFile)
-	json.Unmarshal(byteValue, &CachedAnimeList)
-
-	// Clearing resources
-	defer jsonFile.Close()
-	return true
-}
