+++ src/v1/v1.go
@@ -15,9 +15,9 @@ func Init(r *mux.Router) {
 	v1Route.HandleFunc("/", hdlr.IndexHandler).Methods("GET")
 
 	// Anime list
-	v1Route.HandleFunc("/anime", hdlr.AnimeListHandler).Methods("GET")
-	v1Route.HandleFunc("/anime/", hdlr.AnimeListHandler).Methods("GET")
+	// v1Route.HandleFunc("/anime", hdlr.AnimeListHandler).Methods("GET")
+	// v1Route.HandleFunc("/anime/", hdlr.AnimeListHandler).Methods("GET")
 
-	// Anime by ID
-	v1Route.HandleFunc("/anime/{id:[0-9]+}", hdlr.AnimeHandler).Methods("GET")
+	// // Anime by ID
+	// v1Route.HandleFunc("/anime/{id:[0-9]+}", hdlr.AnimeHandler).Methods("GET")
 }
